refactor(proxy): split bridge account session eviction
