+++ app/modules/proxy/_service/http_bridge/mixin.py
@@ -63,6 +63,7 @@ from app.modules.proxy._service.compact import (
 from app.modules.proxy._service.compact import (
     _sticky_key_from_compact_payload as _sticky_key_from_compact_payload,
 )
+from app.modules.proxy._service.http_bridge.account_sessions import _HTTPBridgeAccountSessionsMixin
 from app.modules.proxy._service.http_bridge.helpers import (
     _active_http_bridge_instance_ring,
     _durable_bridge_lookup_active_owner,
@@ -219,6 +220,7 @@ _HTTP_BRIDGE_BACKGROUND_CLEANUP_WARN_THRESHOLD = 100
 
 class _HTTPBridgeMixin(
     _HTTPBridgeStreamingMixin,
+    _HTTPBridgeAccountSessionsMixin,
     _HTTPBridgeOwnerForwardingMixin,
     _HTTPBridgeRequestSubmitMixin,
     _HTTPBridgeUpstreamEventsMixin,
@@ -358,29 +360,6 @@ class _HTTPBridgeMixin(
                 action="http_bridge_session_close",
                 request_id=_hash_identifier(session.key.affinity_key),
             )
-
-    async def close_http_bridge_sessions_for_account(self, account_id: str) -> int:
-        sessions_to_close: list[_HTTPBridgeSession] = []
-        async with self._http_bridge_lock:
-            for key, session in tuple(self._http_bridge_sessions.items()):
-                if session.account.id != account_id:
-                    continue
-                detached = self._detach_http_bridge_session_locked(key, expected_session=session)
-                if detached is None:
-                    continue
-                _log_http_bridge_event(
-                    "evict_account_binding_changed",
-                    key,
-                    account_id=session.account.id,
-                    model=session.request_model,
-                    cache_key_family=key.affinity_kind,
-                    model_class=_extract_model_class(session.request_model) if session.request_model else None,
-                )
-                sessions_to_close.append(detached)
-
-        for session in sessions_to_close:
-            await self._close_http_bridge_session_bounded(session, reason="account_binding_changed")
-        return len(sessions_to_close)
 
     async def _drain_http_bridge_background_cleanup_tasks(self, *, reason: str) -> None:
         tasks = [
